@@ -9,9 +9,10 @@
 package com.ae.apps.stickerapp;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
+
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class BottomFadingRecyclerView extends RecyclerView {
     public BottomFadingRecyclerView(Context context) {
