@@ -8,11 +8,12 @@
 
 package com.ae.apps.stickerapp;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public class StickerPackListItemViewHolder extends RecyclerView.ViewHolder {
 
