@@ -11,17 +11,24 @@ package com.ae.apps.stickerapp;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.support.annotation.NonNull;
 import android.text.TextUtils;
+import android.util.Log;
+import android.util.Patterns;
+import android.webkit.URLUtil;
+
+import androidx.annotation.NonNull;
 
 import com.facebook.animated.webp.WebPImage;
 
 import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.List;
 
-public class StickerPackValidator {
+class StickerPackValidator {
     private static final int STICKER_FILE_SIZE_LIMIT_KB = 100;
-    private static final int EMOJI_LIMIT = 3;
+    static final int EMOJI_MAX_LIMIT = 3;
+    private static final int EMOJI_MIN_LIMIT = 1;
     private static final int IMAGE_HEIGHT = 512;
     private static final int IMAGE_WIDTH = 512;
     private static final int STICKER_SIZE_MIN = 3;
@@ -31,6 +38,9 @@ public class StickerPackValidator {
     private static final int TRAY_IMAGE_FILE_SIZE_MAX_KB = 50;
     private static final int TRAY_IMAGE_DIMENSION_MIN = 24;
     private static final int TRAY_IMAGE_DIMENSION_MAX = 512;
+    private static final String PLAY_STORE_DOMAIN = "play.google.com";
+    private static final String APPLE_STORE_DOMAIN = "itunes.apple.com";
+
 
     /**
      * Checks whether a sticker pack contains valid data
@@ -58,6 +68,30 @@ public class StickerPackValidator {
         if (TextUtils.isEmpty(stickerPack.trayImageFile)) {
             throw new IllegalStateException("sticker pack tray id is empty, sticker pack identifier:" + stickerPack.identifier);
         }
+        if (!TextUtils.isEmpty(stickerPack.androidPlayStoreLink) && !isValidWebsiteUrl(stickerPack.androidPlayStoreLink)) {
+            throw new IllegalStateException("Make sure to include http or https in url links, android play store link is not a valid url: " + stickerPack.androidPlayStoreLink);
+        }
+        if (!TextUtils.isEmpty(stickerPack.androidPlayStoreLink) && !isURLInCorrectDomain(stickerPack.androidPlayStoreLink, PLAY_STORE_DOMAIN)) {
+            throw new IllegalStateException("android play store link should use play store domain: " + PLAY_STORE_DOMAIN);
+        }
+        if (!TextUtils.isEmpty(stickerPack.iosAppStoreLink) && !isValidWebsiteUrl(stickerPack.iosAppStoreLink)) {
+            throw new IllegalStateException("Make sure to include http or https in url links, ios app store link is not a valid url: " + stickerPack.iosAppStoreLink);
+        }
+        if (!TextUtils.isEmpty(stickerPack.iosAppStoreLink) && !isURLInCorrectDomain(stickerPack.iosAppStoreLink, APPLE_STORE_DOMAIN)) {
+            throw new IllegalStateException("iOS app store link should use app store domain: " + APPLE_STORE_DOMAIN);
+        }
+        if (!TextUtils.isEmpty(stickerPack.licenseAgreementWebsite) && !isValidWebsiteUrl(stickerPack.licenseAgreementWebsite)) {
+            throw new IllegalStateException("Make sure to include http or https in url links, license agreement link is not a valid url: " + stickerPack.licenseAgreementWebsite);
+        }
+        if (!TextUtils.isEmpty(stickerPack.privacyPolicyWebsite) && !isValidWebsiteUrl(stickerPack.privacyPolicyWebsite)) {
+            throw new IllegalStateException("Make sure to include http or https in url links, privacy policy link is not a valid url: " + stickerPack.privacyPolicyWebsite);
+        }
+        if (!TextUtils.isEmpty(stickerPack.publisherWebsite) && !isValidWebsiteUrl(stickerPack.publisherWebsite)) {
+            throw new IllegalStateException("Make sure to include http or https in url links, publisher website link is not a valid url: " + stickerPack.publisherWebsite);
+        }
+        if (!TextUtils.isEmpty(stickerPack.publisherEmail) && !Patterns.EMAIL_ADDRESS.matcher(stickerPack.publisherEmail).matches()) {
+            throw new IllegalStateException("publisher email does not seem valid, email is: " + stickerPack.publisherEmail);
+        }
         try {
             final byte[] bytes = StickerPackLoader.fetchStickerAsset(stickerPack.identifier, stickerPack.trayImageFile, context.getContentResolver());
             if (bytes.length > TRAY_IMAGE_FILE_SIZE_MAX_KB * ONE_KIBIBYTE) {
@@ -83,9 +117,12 @@ public class StickerPackValidator {
     }
 
     private static void validateSticker(@NonNull Context context, @NonNull final String identifier, @NonNull final Sticker sticker) throws IllegalStateException {
-        if (sticker.emojis.size() > EMOJI_LIMIT) {
+        if (sticker.emojis.size() > EMOJI_MAX_LIMIT) {
             throw new IllegalStateException("emoji count exceed limit, sticker pack identifier:" + identifier + ", filename:" + sticker.imageFileName);
         }
+        /* if (sticker.emojis.size() < EMOJI_MIN_LIMIT) {
+            throw new IllegalStateException("To provide best user experience, please associate at least 1 emoji to this sticker, sticker pack identifier:" + identifier + ", filename:" + sticker.imageFileName);
+        }*/
         if (TextUtils.isEmpty(sticker.imageFileName)) {
             throw new IllegalStateException("no file path for sticker, sticker pack identifier:" + identifier);
         }
@@ -99,7 +136,7 @@ public class StickerPackValidator {
                 throw new IllegalStateException("sticker should be less than " + STICKER_FILE_SIZE_LIMIT_KB + "KB, sticker pack identifier:" + identifier + ", filename:" + fileName);
             }
             try {
-                final WebPImage webPImage = WebPImage.create(bytes);
+                final WebPImage webPImage = WebPImage.createFromByteArray(bytes);
                 if (webPImage.getHeight() != IMAGE_HEIGHT) {
                     throw new IllegalStateException("sticker height should be " + IMAGE_HEIGHT + ", sticker pack identifier:" + identifier + ", filename:" + fileName);
                 }
@@ -107,7 +144,7 @@ public class StickerPackValidator {
                     throw new IllegalStateException("sticker width should be " + IMAGE_WIDTH + ", sticker pack identifier:" + identifier + ", filename:" + fileName);
                 }
                 if (webPImage.getFrameCount() > 1) {
-                    throw new IllegalStateException("sticker shoud be a static image, no animated sticker support at the moment, sticker pack identifier:" + identifier + ", filename:" + fileName);
+                    throw new IllegalStateException("sticker should be a static image, no animated sticker support at the moment, sticker pack identifier:" + identifier + ", filename:" + fileName);
                 }
             } catch (IllegalArgumentException e) {
                 throw new IllegalStateException("Error parsing webp image, sticker pack identifier:" + identifier + ", filename:" + fileName, e);
@@ -125,5 +162,31 @@ public class StickerPackValidator {
         if (string.contains("..")) {
             throw new IllegalStateException(string + " cannot contain ..");
         }
+    }
+
+    @SuppressWarnings("BooleanMethodIsAlwaysInverted")
+    private static boolean isValidWebsiteUrl(String websiteUrl) throws IllegalStateException {
+        try {
+            new URL(websiteUrl);
+        } catch (MalformedURLException e) {
+            Log.e("StickerPackValidator", "url: " + websiteUrl + " is malformed");
+            throw new IllegalStateException("url: " + websiteUrl + " is malformed", e);
+        }
+        return URLUtil.isHttpUrl(websiteUrl) || URLUtil.isHttpsUrl(websiteUrl);
+
+    }
+
+    @SuppressWarnings("BooleanMethodIsAlwaysInverted")
+    private static boolean isURLInCorrectDomain(String urlString, String domain) throws IllegalStateException {
+        try {
+            URL url = new URL(urlString);
+            if (domain.equals(url.getHost())) {
+                return true;
+            }
+        } catch (MalformedURLException e) {
+            Log.e("StickerPackValidator", "url: " + urlString + " is malformed");
+            throw new IllegalStateException("url: " + urlString + " is malformed");
+        }
+        return false;
     }
 }
